Move the provider credential routes into their own module
The generic settings helper also owned the payment-provider
credential route factory. This change moves
ProviderCredentialsConfig, persistProviderCredentials, and
defineProviderCredentialsRoute into a focused module,
src/features/admin/settings-provider-credentials.ts. The
Stripe, Square, and SumUp routes now import the factory from
there. TODO.md planned this split, so the entry is deleted.

The mirror test for the factory moves with the code. The Square
settings suite was a direct suite in test/integration/, so the
mutation gate could not select it for its source. It now sits at
its mirror path, test/features/admin/settings-square.test.ts.

No behaviour changes. The moved code is unchanged.
