+++ src/features/admin/settings-square.ts
@@ -7,10 +7,8 @@
 import { settings } from "#db/settings.ts";
 import { t } from "#i18n";
 /* jscpd:ignore-start */
-import {
-  defineProviderCredentialsRoute,
-  settingsSecret,
-} from "#routes/admin/settings-helpers.ts";
+import { settingsSecret } from "#routes/admin/settings-helpers.ts";
+import { defineProviderCredentialsRoute } from "#routes/admin/settings-provider-credentials.ts";
 import { isDemoMode } from "#shared/demo/mode.ts";
 import { squareApi } from "#shared/square/api.ts";
 /* jscpd:ignore-end */
