+++ src/features/admin/settings-sumup.ts
@@ -5,7 +5,7 @@
 
 import { settings } from "#db/settings.ts";
 /* jscpd:ignore-start */
-import { defineProviderCredentialsRoute } from "#routes/admin/settings-helpers.ts";
+import { defineProviderCredentialsRoute } from "#routes/admin/settings-provider-credentials.ts";
 import { isDemoMode } from "#shared/demo/mode.ts";
 import { providerCurrencyBlock } from "#shared/payment-providers.ts";
 import { sumupApi } from "#shared/sumup.ts";
