+++ test/features/admin/settings-provider-credentials.test.ts
@@ -2,7 +2,7 @@ import { expect } from "@std/expect";
 import { fn } from "@std/expect/fn";
 import { it as test } from "@std/testing/bdd";
 import { settings } from "#db/settings.ts";
-import { defineProviderCredentialsRoute } from "#routes/admin/settings-helpers.ts";
+import { defineProviderCredentialsRoute } from "#routes/admin/settings-provider-credentials.ts";
 import { getAllActivityLog } from "#test-utils/activity-log.ts";
 import {
   expectFlash,
